Add version 2.0 after first migration steps.
